More tweaks.
- Dashboard was not including the ticketing steps for non rangers.
- Added vehicle requests to non rangers.
- Prevent radio checkouts on the site check in page if person has not signed the right docs.
