+++ package.json
@@ -80,7 +80,7 @@
     "ember-truth-helpers": "^3.0.0",
     "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-ember": "^11.0.1",
+    "eslint-plugin-ember": "^11.0.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-qunit": "^7.3.1",
     "global": "^4.4.0",
@@ -94,8 +94,8 @@
     "npm-run-all": "^4.1.5",
     "qunit": "^2.19.1",
     "qunit-dom": "^2.0.0",
-    "sass": "^1.53.0",
-    "webpack": "^5.73.0"
+    "sass": "^1.54.0",
+    "webpack": "^5.74.0"
   },
   "engines": {
     "node": "14.* || >= 16"
